Refactor to remove container dir
