+++ src/components/organisms/CalendarList.tsx
@@ -1,19 +1,19 @@
 import React from "react";
 import styled from "styled-components";
-import { UserCalendarContainer } from "../../containers/UserCalendarContainer";
-import { TemplateCalendarContainer } from "../../containers/TemplateCalendarContainer";
+import { UserCalendarContainer } from "../molecules/UserCalendarContainer";
+import { TemplateCalendarContainer } from "../molecules/TemplateCalendarContainer";
 
 export function CalendarList() {
   return (
     <Container>
       <div>
-        <h2>UserCalendar</h2>
+        <h2>User Calendar</h2>
         <UserCalendarContainer />
       </div>
 
       <hr />
       <div>
-        <h2>CommunityCollege</h2>
+        <h2>Community College</h2>
         <TemplateCalendarContainer />
       </div>
     </Container>
